refactor: add SearchSettings component
